Audio Cookie fixes + Choice menu vanishing.
Audio Cookie fixes

User's who put audio volumes to 0 nd reloaded the adventure page had their audio set to the default.
The cause was cookies set to 0 turns into a null, causing it to default.

Resolved this by adding a new cookie to see if the user was new or not.
If they are, the volumes are defaulted, if they aren't then it proceeds to use the cookies.

Choice node had an issue where the input window wouldn't go away. I didnt add in a check to set the choice menu visibility to false when an input was made.